Performance Measurements Changes
So we've ditched the testing accuracy measurement cos I'm not it was right. Instead on the testing accuracy we use balanced accuracy score because it helps measure accuracy better on datasets that aren't perfectly balanced to each class. We now have all classifiers running under 10 seconds with the neural network the most accurate, as expected