Minor Changes
Retuned the random_state params on some of the classifiers and tweaked some of the grid search params. They give more realistic results which I think could probably be tuned to be better given more time to play around with it but for now it does its job quite well. Weirdly the nearest neighbours classifier is the least accurate, and the decision tree is more accurate. Probably worth noting that in the report.